requirements.txt
pandas<=1.3.5
catboost<=1.2.2
grpcio<=1.60.1
grpcio-tools<=1.60.1
protobuf<=4.25.3
scikit-learn<=1.5.0
requests<=2.27.1
pydantic<=2.7.1
